automagic/forms.py
from django import forms
from django.utils.translation import ugettext_lazy as _

from crispy_forms.helper import FormHelper
from crispy_forms.layout import Submit


class AutomagicForm(forms.ModelForm):

    def __init__(self, *args, **kwargs):
        super(AutomagicForm, self).__init__(*args, **kwargs)
        self.helper = FormHelper()
        self.helper.add_input(Submit('submit', _('Submit')))
