requirements.txt
Django>=1.7,<=1.8
django-braces==1.4.0
django-crispy-forms==1.4.0